Transport can be abstracted for mocking.
This is a stateless request sending interface that can be replaced for a fake in
tests.

The only real change is that the responsibility for generating headers is now in
the client instead of the transport so the network specific stuff doesn't need
to know about access tokens any more.