Use Psr\Log from Composer, add phpunit.xml and bootstrap
This commit removes a duplicated version of Psr\Log, which could have
caused applications to use a different version than expected. Instead,
we list psr/log as a composer dependency.

To facilitate unit tests, a phpunit-bootstrap file is added which
includes vendor/autoload.php, and a phpunit.xml file is added to point
to the bootstrap.

Running unit tests now requires 'composer install', but afterward you
simply run 'phpunit' in the root directory of the project.
'composer test' also works.

Phpunit is also added to the require-dev section of composer.json, as
the unit tests do depend on it.

Fixes #58